pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = ""
authors = [
    {name = "RWallan",email = "3am.richardwallan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "notebook (>=7.4.2,<8.0.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "duckduckgo-search (==5.3.1b1)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "streamlit (>=1.45.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
